requirements.txt
SQLAlchemy==2.0.43
requests
python-dotenv
